+++ sw.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = 'meteoasturlode-v205-radarmarkupclean';
+const CACHE_NAME = 'meteoasturlode-v206-foehn';
 const STATIC_ASSETS = [
   './',
   './index.html',
@@ -33,6 +33,7 @@ const STATIC_ASSETS = [
   './js/utils/weatherAlerts.js',
   './js/utils/weatherExplanations.js',
   './js/utils/weatherPhenomena.js',
+  './js/utils/foehnDetector.js',
   './js/utils/storage.js',
   './js/utils/tides.js'
 ];